// frontend/src/api/generated/shop/schemas/orderProcessingNote.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * ShopDjango API
 * API for ShopDjango project
 * OpenAPI spec version: 1.0.0
 */

export interface OrderProcessingNote {
  readonly id: number;
  /** Order this note relates to */
  order: number;
  /** Staff member who created this note */
  staff_member: number;
  /** Note content */
  note: string;
  /** Timestamp when the record was created */
  readonly created_at: string;
  /** Timestamp when the record was last updated */
  readonly updated_at: string;
}